finish 1st review round
mainly rm unused code and dangling comments
