+++ test/python/circuit/test_parameters.py
@@ -1599,8 +1599,7 @@ class TestParameterExpressions(QiskitTestCase):
         with self.assertRaisesRegex(CircuitError, "Name conflict"):
             expr.subs({x: y_second})
 
-    # @data(2, 1.3, 0, -1, -1.0, numpy.pi, 1j)
-    @data(0)
+    @data(2, 1.3, 0, -1, -1.0, numpy.pi, 1j)
     def test_expressions_of_parameter_with_constant(self, const):
         """Verify operating on a Parameter with a constant."""
 
@@ -2191,15 +2190,6 @@ class TestParameterEquality(QiskitTestCase):
 
         self.assertEqual(expr, theta)
         self.assertEqual(theta, expr)
-
-    # Julien: this test was using the internal ._name_map property -- rewrite using
-    # actual public interface or delete this test
-    # def test_parameter_symbol_equal_after_ufunc(self):
-    #     """Verify ParameterExpression phi
-    #     and ParameterExpression cos(phi) have the same symbol map"""
-    #     phi = Parameter("phi")
-    #     cos_phi = numpy.cos(phi)
-    #     self.assertEqual(phi.name_map(), cos_phi.name_map())
 
 
 class TestParameterView(QiskitTestCase):
